Refactor mechanism list
